// template/react-starter-kit/src/.generated/api/model/supportTicketItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */
import type { SupportTicketPriority } from './supportTicketPriority';
import type { SupportTicketStatus } from './supportTicketStatus';

export interface SupportTicketItemDto {
  id: string;
  userId: string;
  userName: string;
  /** @nullable */
  assigneeId: string | null;
  /** @nullable */
  assigneeName: string | null;
  category: string;
  title: string;
  content: string;
  priority: SupportTicketPriority;
  status: SupportTicketStatus;
  /** @nullable */
  resolution: string | null;
  createdAt: string;
  updatedAt: string;
}